chore: enable PMD on Platform code (#13172)
Enable PMD on platform code by bumping the plugin versions to the versions where PMD enabled.

This is a big PR because of several new PMD rules introduced,
- UnnecessaryBoxing.
- LambdaCanBeMethodReference
- UnnecessaryVarargsArrayCreation
- ConfusingArgumentToVarargsMethod
- UseEnumCollections

Wherever possible I fixed as much as I could.
- I was a bit more lax with test classes, namely, I was liberal with AvoidDuplicateLiterals suppression. Sometimes it's useful to not DRY test variables so changes are more intentional.
- There are several places where we use try-with-resources to properly close a resources, while not using the resource itself e.g. logging. In those cases, I suppressed the UnusedLocalVariable warning.
- I did not fix any DB migration code. Since migrations are meant to be immutable, I felt it was clearer to leave them as such.
- There were a surprising amount of dead code/unused variables. These I removed.

Overall the changes, while widespread, are minor. The only places I'm worried about are the authentication code. I'll run them by the Compose OC before merging in.

+++ airbyte-data/src/main/java/io/airbyte/data/services/impls/jooq/DbConverter.java
@@ -253,7 +253,7 @@ public class DbConverter {
   public static StandardSourceDefinition buildStandardSourceDefinition(final Record record, final long defaultMaxSecondsBetweenMessages) {
     var maxSecondsBetweenMessage = record.get(ACTOR_DEFINITION.MAX_SECONDS_BETWEEN_MESSAGES) == null
         ? defaultMaxSecondsBetweenMessages
-        : record.get(ACTOR_DEFINITION.MAX_SECONDS_BETWEEN_MESSAGES).longValue();
+        : record.get(ACTOR_DEFINITION.MAX_SECONDS_BETWEEN_MESSAGES);
 
     // All sources are starting to set this field according to their rate limits. As a
     // safeguard for sources with rate limits that are too low e.g. minutes etc, we default to
